move refresh to correct filepath
